Corrected DTO and updated md file.

+++ Backlog.Api/DTOs/BacklogDtos.cs
@@ -3,4 +3,4 @@ namespace Backlog.Api.Dtos;
 
 public record BacklogItemDto(Guid Id, string Title, Category Category, BacklogStatus Status, DateTime CreatedAt, decimal? Rating, String? Note);
 public record CreateBacklogItemRequest(string Title, Category Category);
-public record UpdateBacklogItemRequest(BacklogStatus Status, int? Rating, String? Note);
+public record UpdateBacklogItemRequest(BacklogStatus Status, decimal? Rating, String? Note);
